0.0.0-replace-float-with-double.1+1
There is no performance gain when using floating points, since matrices are multiplied on CPU.

+++ app/src/main/java/io/jim/tesserapp/geometry/Grid.kt
@@ -14,19 +14,19 @@ class Grid(
 
     init {
         for (i in -5..5) {
-            addLine(Vector4dh(i.toFloat(), 0f, -5f, 0f), Vector4dh(i.toFloat(), 0f, 5f, 0f))
-            addLine(Vector4dh(-5f, 0f, i.toFloat(), 0f), Vector4dh(5f, 0f, i.toFloat(), 0f))
+            addLine(Vector4dh(i.toDouble(), 0.0, -5.0, 0.0), Vector4dh(i.toDouble(), 0.0, 5.0, 0.0))
+            addLine(Vector4dh(-5.0, 0.0, i.toDouble(), 0.0), Vector4dh(5.0, 0.0, i.toDouble(), 0.0))
         }
         for (i in 1..5) {
-            addLine(Vector4dh(i.toFloat(), 0f, -5f, 0f), Vector4dh(i.toFloat(), 0f, 5f, 0f))
-            addLine(Vector4dh(-5f, 0f, i.toFloat(), 0f), Vector4dh(5f, 0f, i.toFloat(), 0f))
+            addLine(Vector4dh(i.toDouble(), 0.0, -5.0, 0.0), Vector4dh(i.toDouble(), 0.0, 5.0, 0.0))
+            addLine(Vector4dh(-5.0, 0.0, i.toDouble(), 0.0), Vector4dh(5.0, 0.0, i.toDouble(), 0.0))
         }
 
-        addLine(Vector4dh(-5f, 0f, 0f, 0f), Vector4dh(0f, 0f, 0f, 0f))
-        addLine(Vector4dh(1f, 0f, 0f, 0f), Vector4dh(5f, 0f, 0f, 0f))
+        addLine(Vector4dh(-5.0, 0.0, 0.0, 0.0), Vector4dh(0.0, 0.0, 0.0, 0.0))
+        addLine(Vector4dh(1.0, 0.0, 0.0, 0.0), Vector4dh(5.0, 0.0, 0.0, 0.0))
 
-        addLine(Vector4dh(0f, 0f, -5f, 0f), Vector4dh(0f, 0f, 0f, 0f))
-        addLine(Vector4dh(0f, 0f, 1f, 0f), Vector4dh(0f, 0f, 5f, 0f))
+        addLine(Vector4dh(0.0, 0.0, -5.0, 0.0), Vector4dh(0.0, 0.0, 0.0, 0.0))
+        addLine(Vector4dh(0.0, 0.0, 1.0, 0.0), Vector4dh(0.0, 0.0, 5.0, 0.0))
     }
 
 }
