+++ app/src/main/java/io/jim/tesserapp/math/common/constants.kt
@@ -3,4 +3,4 @@ package io.jim.tesserapp.math.common
 /**
  * Pi, in floating-point precision.
  */
-const val Pi: Float = 3.14159265358979323846f
+const val Pi: Double = 3.14159265358979323846
