+++ app/src/main/java/io/jim/tesserapp/ui/CubeView.kt
@@ -35,7 +35,7 @@ class CubeView : FrameLayout {
             /**
              * Rotate the cube around the x-axis.
              */
-            override var x = 0f
+            override var x = 0.0
                 set(value) {
                     coordinateSystemView.sharedRenderData.synchronized {
                         cube.smoothRotation.x = value
@@ -45,7 +45,7 @@ class CubeView : FrameLayout {
             /**
              * Rotate the cube around the y-axis.
              */
-            override var y = 0f
+            override var y = 0.0
                 set(value) {
                     coordinateSystemView.sharedRenderData.synchronized {
                         cube.smoothRotation.y = value
@@ -55,7 +55,7 @@ class CubeView : FrameLayout {
             /**
              * Rotate the cube around the z-axis.
              */
-            override var z = 0f
+            override var z = 0.0
                 set(value) {
                     coordinateSystemView.sharedRenderData.synchronized {
                         cube.smoothRotation.z = value
@@ -65,7 +65,7 @@ class CubeView : FrameLayout {
             /**
              *  Rotate the cube around on the q-x plane.
              */
-            override var q = 0f
+            override var q = 0.0
                 set(value) {
                     coordinateSystemView.sharedRenderData.synchronized {
                         //cube.rotation.q = value
@@ -79,7 +79,7 @@ class CubeView : FrameLayout {
             /**
              * Translates the cube along the x-axis.
              */
-            override var x = 0f
+            override var x = 0.0
                 set(value) {
                     coordinateSystemView.sharedRenderData.synchronized {
                         cube.smoothTranslation.x = value
@@ -89,7 +89,7 @@ class CubeView : FrameLayout {
             /**
              * Translates the cube along the y-axis.
              */
-            override var y = 0f
+            override var y = 0.0
                 set(value) {
                     coordinateSystemView.sharedRenderData.synchronized {
                         cube.smoothTranslation.y = value
@@ -99,7 +99,7 @@ class CubeView : FrameLayout {
             /**
              * Translates the cube along the z-axis.
              */
-            override var z = 0f
+            override var z = 0.0
                 set(value) {
                     coordinateSystemView.sharedRenderData.synchronized {
                         cube.smoothTranslation.z = value
@@ -109,7 +109,7 @@ class CubeView : FrameLayout {
             /**
              * Translates the cube along the q-axis.
              */
-            override var q = 0f
+            override var q = 0.0
                 set(value) {
                     coordinateSystemView.sharedRenderData.synchronized {
                         cube.smoothTranslation.q = value
@@ -121,7 +121,7 @@ class CubeView : FrameLayout {
         /**
          * Control the camera distance.
          */
-        override var cameraDistance: Float by Delegates.observable(0f) { _, _, newValue ->
+        override var cameraDistance: Double by Delegates.observable(0.0) { _, _, newValue ->
             coordinateSystemView.sharedRenderData.synchronized {
                 coordinateSystemView.sharedRenderData.camera.distance = newValue
             }
@@ -144,13 +144,13 @@ class CubeView : FrameLayout {
      */
     private val cube = Geometry("Cube", Color(context, R.color.colorAccent)).apply {
         addQuadrilateral(
-                Vector4dh(1f, 1f, 1f, 0f),
-                Vector4dh(-1f, 1f, 1f, 0f),
-                Vector4dh(-1f, -1f, 1f, 0f),
-                Vector4dh(1f, -1f, 1f, 0f)
+                Vector4dh(1.0, 1.0, 1.0, 0.0),
+                Vector4dh(-1.0, 1.0, 1.0, 0.0),
+                Vector4dh(-1.0, -1.0, 1.0, 0.0),
+                Vector4dh(1.0, -1.0, 1.0, 0.0)
         )
 
-        extrude(Vector4dh(0f, 0f, -2f, 0f))
+        extrude(Vector4dh(0.0, 0.0, -2.0, 0.0))
 
         coordinateSystemView.sharedRenderData.drawDataProvider += this
     }
